Separate Pools to Pool class
